+++ src/react/theme/index.js
@@ -1,2 +1,3 @@
 export { EFFECT } from './constants';
+export { setGlobalStyle } from './globalStyles';
 export { getColorEffect, getTextColor, getHeaderTheme, getNavigationTheme } from './ThemeHelper';