+++ api/CHANGELOG.md
@@ -9,6 +9,20 @@ Some extra tags:
 - [rpc] tag indicates that the change affects RPC API
 - [breaking] tag indicates that the change is not backwards compatible
 
+## 2.10.0
+
+### [rem] MapResizeEnabledRequest
+
+Removed request that was not used or needed anymore. The request could be used for stopping mapmodule from auto-scaling its size when the element size changed, but the current implementation works in a similar way that you can set size of the map element to be different than the container it is in. This does the same thing.
+
+### [rem] MapSizeUpdateRequest
+
+Removed request that was not used or needed anymore now that mapmodule listens to the element size. The request was NOT used to set a new size like the name suggests, but to notify mapmodule about size change like it was an event. This should have been an event instead, but it's no longer required.
+
+### [rem] MapWindowFullScreenRequest
+
+Removed request that was not needed anymore now that mapmodule listens to the element size. The request was NOT used to set a new size for map like the name suggests, but to notify mapmodule about size change like it was an event. This should have been an event instead, but it's no longer required.
+
 ## 2.9.0
 
 ### [mod] [rpc] [breaking] MapModulePlugin.AddMarkerRequest